requirements.txt
appdirs==1.4.3
appnope==0.1.0
arrow==0.10.0
backcall==0.1.0
bleach==3.0.2
cachetools==3.0.0
decorator==4.3.0
defusedxml==0.5.0
Django==1.10.4
django-cors-headers==2.4.0
django-crontab==0.7.1
django-extensions==1.7.5
django-filter==1.0.1
djangorestframework==3.5.3
entrypoints==0.2.3
google-api-python-client==1.7.4
google-auth==1.6.0
google-auth-httplib2==0.0.3
html5lib==0.9999999
httplib2==0.11.3
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Markdown==2.6.7
MarkupSafe==1.1.0
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.0
oauth2client==4.1.3
packaging==16.8
pandocfilters==1.4.2
parse-rest==0.2.20141004
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.4.2
prompt-toolkit==2.0.7
psycopg2==2.7.6.1
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.2.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2016.10
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
rsa==4.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
uritemplate==3.0.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.11.11
widgetsnbextension==3.4.2
